@david-xl 2023-10-09
Curious: why is benchmark 3's runtime increasing (marked as benchmark 2 in Before)?