@spupyrev 2023-10-09
> Curious: why is benchmark 3's runtime increasing (marked as benchmark 2 in Before)?

I wasn't particularly careful with measuring build times here, as these are seconds, while the overall compilation takes minutes/hours. We should treat the "reordering runtime" here as given with a +/- 10% precision, so the difference for "benchmark 3" (sorry for the typo) is likely just noise. In my mind, the times for benchmarks 1,2,3 are unchanged, while there is small but measurable (~1 sec) improvement for benchmark 4, and a large improvement for benchmark 5.
Let me know if you'd like to get higher precision.