@spupyrev 2023-10-13
After more thorough testing, I had to bring back `ChainSplitThreshold` and modify the tests accordingly as otherwise some tests were up to 4x slower. Having this threshold regresses the quality a little bit, but I don't see a clean way.
Adjusted the numbers on my benchmarks to reflect the change.